requirements.txt
Flask
gunicorn
psycopg2-binary
python-dotenv
line-bot-sdk
supabase-py
